+++ src/changelog.md
@@ -1,4 +1,3 @@
 ## changelog
-- increase http timeout time for getting charge current values 
-- substitute charge current values in case either one fails to respond
-- fix incorrect handling of data retrieving commands
+- add voltage settings
+- upgrade packages